requirements.txt
django-heroku==0.3.1
Django==4.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
mixer==7.2.2
psycopg2==2.9.5
python-dotenv==0.21.0
whitenoise==6.2.0
